adding Lmk04832 python support
