+++ python/surf/devices/ti/__init__.py
@@ -16,7 +16,9 @@ from surf.devices.ti._Ads54J60Channel   import *
 from surf.devices.ti._AxiCdcm6208       import *
 from surf.devices.ti._Dac38J84          import *
 from surf.devices.ti._Ds32Ev400         import *
+from surf.devices.ti._Lmk048Base        import *
 from surf.devices.ti._Lmk04828          import *
+from surf.devices.ti._Lmk04832          import *
 from surf.devices.ti._Lmk61e2           import *
 from surf.devices.ti._Pca9535           import *
 from surf.devices.ti._UCD92xx           import *
